requirements.txt
asgiref==3.2.3
autopep8==1.4.4
Babel==2.8.0
boto3==1.11.14
botocore==1.14.14
certifi==2019.11.28
chardet==3.0.4
defusedxml==0.6.0
diff-match-patch==20181111
Django==3.0.3
django-cors-headers==3.2.1
django-environ==0.4.5
django-extensions==2.2.8
django-filter==2.2.0
django-import-export==2.0.1
django-phonenumber-field==4.0.0
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
djangorestframework-xml==1.4.0
docutils==0.15.2
et-xmlfile==1.0.1
idna==2.8
jdcal==1.4.1
jmespath==0.9.4
MarkupPy==1.14
numpy==1.18.1
odfpy==1.4.1
openpyxl==3.0.3
pandas==1.0.1
phonenumbers==8.11.3
pipenv==2018.11.26
psycopg2-binary==2.8.4
pycodestyle==2.5.0
PyJWT==1.7.1
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.3
requests==2.22.0
s3transfer==0.3.3
six==1.14.0
sqlparse==0.3.0
tablib==0.14.0
urllib3==1.25.8
virtualenv==16.7.9
virtualenv-clone==0.5.3
xlrd==1.2.0
xlwt==1.3.0
